server_py/requirements.txt
kokoro>=0.9.4
soundfile
grpcio
grpcio-tools
numpy
